Day 23 - Added rate limiting, brute force protection and OTP security

+++ src/main/java/com/smartwallet/service/OtpService.java
@@ -1,8 +1,10 @@
 package com.smartwallet.service;
 
 import com.smartwallet.dto.VerifyOtpRequest;
+import com.smartwallet.exception.TooManyRequestsException;
 import com.smartwallet.model.Otp;
 import com.smartwallet.repository.OtpRepository;
+import com.smartwallet.security.RateLimiterService;
 import java.time.LocalDateTime;
 import java.util.Random;
 import org.springframework.stereotype.Service;
@@ -12,25 +14,35 @@ public class OtpService {
 
     private final OtpRepository otpRepository;
     private final EmailService emailService;
+    private final RateLimiterService rateLimiterService;
 
     public OtpService(
             OtpRepository otpRepository,
-            EmailService emailService) {
+            EmailService emailService,
+            RateLimiterService rateLimiterService) {
 
         this.otpRepository = otpRepository;
         this.emailService = emailService;
+        this.rateLimiterService = rateLimiterService;
     }
 
     public String sendOtp(String email) {
-        ///if user register then this method call and generate Otp
+
+        if (!rateLimiterService.isAllowed(email)) {
+
+            throw new TooManyRequestsException(
+                    "Too many OTP requests. Try again after 1 minute."
+            );
+        }
+
         String otp =
                 String.valueOf(
                         100000 + new Random().nextInt(900000)
                 );
-       //Check Existing OTP
+
         Otp existingOtp =
                 otpRepository.findByEmail(email);
-       //Delete Old OTP
+
         if (existingOtp != null) {
             otpRepository.delete(existingOtp);
         }
@@ -44,7 +56,10 @@ public class OtpService {
 
         otpRepository.save(otpEntity);
 
-        emailService.sendOtpEmail(email, otp);
+        emailService.sendOtpEmail(
+                email,
+                otp
+        );
 
         return "OTP sent successfully";
     }
